requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
secure-smtplib